Refactoring KafkaIndexTask for better code maintenance
