+++ extensions-core/kafka-indexing-service/src/main/java/io/druid/indexing/kafka/LegacyKafkaIndexTaskRunner.java
@@ -1,0 +1,1164 @@
+/*
+ * Licensed to Metamarkets Group Inc. (Metamarkets) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. Metamarkets licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package io.druid.indexing.kafka;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.base.Function;
+import com.google.common.base.Joiner;
+import com.google.common.base.Optional;
+import com.google.common.base.Preconditions;
+import com.google.common.base.Supplier;
+import com.google.common.base.Throwables;
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Iterables;
+import com.google.common.collect.Maps;
+import com.google.common.collect.Sets;
+import io.druid.data.input.Committer;
+import io.druid.data.input.InputRow;
+import io.druid.data.input.impl.InputRowParser;
+import io.druid.discovery.DiscoveryDruidNode;
+import io.druid.discovery.DruidNodeDiscoveryProvider;
+import io.druid.discovery.LookupNodeService;
+import io.druid.indexer.IngestionState;
+import io.druid.indexer.TaskMetricsGetter;
+import io.druid.indexing.common.IngestionStatsAndErrorsTaskReport;
+import io.druid.indexing.common.IngestionStatsAndErrorsTaskReportData;
+import io.druid.indexing.common.TaskRealtimeMetricsMonitorBuilder;
+import io.druid.indexing.common.TaskReport;
+import io.druid.indexing.common.TaskStatus;
+import io.druid.indexing.common.TaskToolbox;
+import io.druid.indexing.common.actions.ResetDataSourceMetadataAction;
+import io.druid.indexing.common.actions.SegmentTransactionalInsertAction;
+import io.druid.indexing.common.task.IndexTaskUtils;
+import io.druid.indexing.common.task.RealtimeIndexTask;
+import io.druid.indexing.kafka.KafkaIndexTask.Status;
+import io.druid.java.util.common.DateTimes;
+import io.druid.java.util.common.ISE;
+import io.druid.java.util.common.StringUtils;
+import io.druid.java.util.common.collect.Utils;
+import io.druid.java.util.common.parsers.ParseException;
+import io.druid.java.util.emitter.EmittingLogger;
+import io.druid.segment.indexing.RealtimeIOConfig;
+import io.druid.segment.realtime.FireDepartment;
+import io.druid.segment.realtime.FireDepartmentMetrics;
+import io.druid.segment.realtime.FireDepartmentMetricsTaskMetricsGetter;
+import io.druid.segment.realtime.appenderator.Appenderator;
+import io.druid.segment.realtime.appenderator.AppenderatorDriverAddResult;
+import io.druid.segment.realtime.appenderator.SegmentIdentifier;
+import io.druid.segment.realtime.appenderator.SegmentsAndMetadata;
+import io.druid.segment.realtime.appenderator.StreamAppenderatorDriver;
+import io.druid.segment.realtime.appenderator.TransactionalSegmentPublisher;
+import io.druid.segment.realtime.firehose.ChatHandlerProvider;
+import io.druid.server.security.Access;
+import io.druid.server.security.Action;
+import io.druid.server.security.AuthorizerMapper;
+import io.druid.timeline.DataSegment;
+import io.druid.utils.CircularBuffer;
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.kafka.clients.consumer.ConsumerRecords;
+import org.apache.kafka.clients.consumer.KafkaConsumer;
+import org.apache.kafka.clients.consumer.OffsetOutOfRangeException;
+import org.apache.kafka.common.TopicPartition;
+import org.joda.time.DateTime;
+
+import javax.annotation.Nullable;
+import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.DefaultValue;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+import java.io.IOException;
+import java.nio.ByteBuffer;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.Future;
+import java.util.concurrent.RejectedExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.locks.Condition;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
+import java.util.stream.Collectors;
+
+/**
+ * Kafka index task runner which doesn't support incremental segment publishing. We keep this to support rolling update.
+ * This class will be removed in a future release.
+ */
+@Deprecated
+public class LegacyKafkaIndexTaskRunner implements KafkaIndexTaskRunner
+{
+  private static final EmittingLogger log = new EmittingLogger(LegacyKafkaIndexTaskRunner.class);
+  private static final String METADATA_NEXT_PARTITIONS = "nextPartitions";
+
+  private final Map<Integer, Long> endOffsets = new ConcurrentHashMap<>();
+  private final Map<Integer, Long> nextOffsets = new ConcurrentHashMap<>();
+
+  // The pause lock and associated conditions are to support coordination between the Jetty threads and the main
+  // ingestion loop. The goal is to provide callers of the API a guarantee that if pause() returns successfully
+  // the ingestion loop has been stopped at the returned offsets and will not ingest any more data until resumed. The
+  // fields are used as follows (every step requires acquiring [pauseLock]):
+  //   Pausing:
+  //   - In pause(), [pauseRequested] is set to true and then execution waits for [status] to change to PAUSED, with the
+  //     condition checked when [hasPaused] is signalled.
+  //   - In possiblyPause() called from the main loop, if [pauseRequested] is true, [status] is set to PAUSED,
+  //     [hasPaused] is signalled, and execution pauses until [pauseRequested] becomes false, either by being set or by
+  //     the [pauseMillis] timeout elapsing. [pauseRequested] is checked when [shouldResume] is signalled.
+  //   Resuming:
+  //   - In resume(), [pauseRequested] is set to false, [shouldResume] is signalled, and execution waits for [status] to
+  //     change to something other than PAUSED, with the condition checked when [shouldResume] is signalled.
+  //   - In possiblyPause(), when [shouldResume] is signalled, if [pauseRequested] has become false the pause loop ends,
+  //     [status] is changed to STARTING and [shouldResume] is signalled.
+
+  private final Lock pauseLock = new ReentrantLock();
+  private final Condition hasPaused = pauseLock.newCondition();
+  private final Condition shouldResume = pauseLock.newCondition();
+
+  private final AtomicBoolean stopRequested = new AtomicBoolean(false);
+  private final AtomicBoolean publishOnStop = new AtomicBoolean(false);
+
+  // [statusLock] is used to synchronize the Jetty thread calling stopGracefully() with the main run thread. It prevents
+  // the main run thread from switching into a publishing state while the stopGracefully() thread thinks it's still in
+  // a pre-publishing state. This is important because stopGracefully() will try to use the [stopRequested] flag to stop
+  // the main thread where possible, but this flag is not honored once publishing has begun so in this case we must
+  // interrupt the thread. The lock ensures that if the run thread is about to transition into publishing state, it
+  // blocks until after stopGracefully() has set [stopRequested] and then does a final check on [stopRequested] before
+  // transitioning to publishing state.
+  private final Object statusLock = new Object();
+
+  private final Lock pollRetryLock = new ReentrantLock();
+  private final Condition isAwaitingRetry = pollRetryLock.newCondition();
+
+  private final KafkaIndexTask task;
+  private final KafkaIOConfig ioConfig;
+  private final KafkaTuningConfig tuningConfig;
+  private final InputRowParser<ByteBuffer> parser;
+  private final AuthorizerMapper authorizerMapper;
+  private final Optional<ChatHandlerProvider> chatHandlerProvider;
+  private final CircularBuffer<Throwable> savedParseExceptions;
+
+  private volatile DateTime startTime;
+  private volatile Status status = Status.NOT_STARTED; // this is only ever set by the task runner thread (runThread)
+  private volatile ObjectMapper objectMapper;
+  private volatile Thread runThread;
+  private volatile Appenderator appenderator;
+  private volatile StreamAppenderatorDriver driver;
+  private volatile FireDepartmentMetrics fireDepartmentMetrics;
+  private volatile TaskMetricsGetter metricsGetter;
+  private volatile IngestionState ingestionState;
+
+  private volatile long pauseMillis = 0;
+  private volatile boolean pauseRequested;
+
+  LegacyKafkaIndexTaskRunner(
+      KafkaIndexTask task,
+      InputRowParser<ByteBuffer> parser,
+      AuthorizerMapper authorizerMapper,
+      Optional<ChatHandlerProvider> chatHandlerProvider,
+      CircularBuffer<Throwable> savedParseExceptions
+  )
+  {
+    this.task = task;
+    this.ioConfig = task.getIOConfig();
+    this.tuningConfig = task.getTuningConfig();
+    this.parser = parser;
+    this.authorizerMapper = authorizerMapper;
+    this.chatHandlerProvider = chatHandlerProvider;
+    this.savedParseExceptions = savedParseExceptions;
+
+    this.endOffsets.putAll(ioConfig.getEndPartitions().getPartitionOffsetMap());
+    this.ingestionState = IngestionState.NOT_STARTED;
+  }
+
+  @Override
+  public TaskStatus run(TaskToolbox toolbox)
+  {
+    try {
+      return runInternal(toolbox);
+    }
+    catch (Exception e) {
+      log.error(e, "Encountered exception while running task.");
+      final String errorMsg = Throwables.getStackTraceAsString(e);
+      toolbox.getTaskReportFileWriter().write(getTaskCompletionReports(errorMsg));
+      return TaskStatus.failure(
+          task.getId(),
+          errorMsg
+      );
+    }
+  }
+
+  @Override
+  public Appenderator getAppenderator()
+  {
+    return appenderator;
+  }
+
+  @Override
+  public FireDepartmentMetrics getFireDepartmentMetrics()
+  {
+    return fireDepartmentMetrics;
+  }
+
+  private TaskStatus runInternal(TaskToolbox toolbox) throws Exception
+  {
+    log.info("Starting up!");
+    startTime = DateTimes.nowUtc();
+    status = Status.STARTING;
+    objectMapper = toolbox.getObjectMapper();
+
+    if (chatHandlerProvider.isPresent()) {
+      log.info("Found chat handler of class[%s]", chatHandlerProvider.get().getClass().getName());
+      chatHandlerProvider.get().register(task.getId(), this, false);
+    } else {
+      log.warn("No chat handler detected");
+    }
+
+    runThread = Thread.currentThread();
+
+    // Set up FireDepartmentMetrics
+    final FireDepartment fireDepartmentForMetrics = new FireDepartment(
+        task.getDataSchema(),
+        new RealtimeIOConfig(null, null, null),
+        null
+    );
+    fireDepartmentMetrics = fireDepartmentForMetrics.getMetrics();
+    metricsGetter = new FireDepartmentMetricsTaskMetricsGetter(fireDepartmentMetrics);
+    toolbox.getMonitorScheduler().addMonitor(TaskRealtimeMetricsMonitorBuilder.build(task, fireDepartmentForMetrics));
+
+    final String lookupTier = task.getContextValue(RealtimeIndexTask.CTX_KEY_LOOKUP_TIER);
+    LookupNodeService lookupNodeService = lookupTier == null ?
+                                          toolbox.getLookupNodeService() :
+                                          new LookupNodeService(lookupTier);
+    DiscoveryDruidNode discoveryDruidNode = new DiscoveryDruidNode(
+        toolbox.getDruidNode(),
+        DruidNodeDiscoveryProvider.NODE_TYPE_PEON,
+        ImmutableMap.of(
+            toolbox.getDataNodeService().getName(), toolbox.getDataNodeService(),
+            lookupNodeService.getName(), lookupNodeService
+        )
+    );
+
+    ingestionState = IngestionState.BUILD_SEGMENTS;
+
+    try (
+        final Appenderator appenderator0 = task.newAppenderator(fireDepartmentMetrics, toolbox);
+        final StreamAppenderatorDriver driver = task.newDriver(appenderator0, toolbox, fireDepartmentMetrics);
+        final KafkaConsumer<byte[], byte[]> consumer = task.newConsumer()
+    ) {
+      toolbox.getDataSegmentServerAnnouncer().announce();
+      toolbox.getDruidNodeAnnouncer().announce(discoveryDruidNode);
+
+      appenderator = appenderator0;
+
+      final String topic = ioConfig.getStartPartitions().getTopic();
+
+      // Start up, set up initial offsets.
+      final Object restoredMetadata = driver.startJob();
+      if (restoredMetadata == null) {
+        nextOffsets.putAll(ioConfig.getStartPartitions().getPartitionOffsetMap());
+      } else {
+        final Map<String, Object> restoredMetadataMap = (Map) restoredMetadata;
+        final KafkaPartitions restoredNextPartitions = toolbox.getObjectMapper().convertValue(
+            restoredMetadataMap.get(METADATA_NEXT_PARTITIONS),
+            KafkaPartitions.class
+        );
+        nextOffsets.putAll(restoredNextPartitions.getPartitionOffsetMap());
+
+        // Sanity checks.
+        if (!restoredNextPartitions.getTopic().equals(ioConfig.getStartPartitions().getTopic())) {
+          throw new ISE(
+              "WTF?! Restored topic[%s] but expected topic[%s]",
+              restoredNextPartitions.getTopic(),
+              ioConfig.getStartPartitions().getTopic()
+          );
+        }
+
+        if (!nextOffsets.keySet().equals(ioConfig.getStartPartitions().getPartitionOffsetMap().keySet())) {
+          throw new ISE(
+              "WTF?! Restored partitions[%s] but expected partitions[%s]",
+              nextOffsets.keySet(),
+              ioConfig.getStartPartitions().getPartitionOffsetMap().keySet()
+          );
+        }
+      }
+
+      // Set up sequenceNames.
+      final Map<Integer, String> sequenceNames = Maps.newHashMap();
+      for (Integer partitionNum : nextOffsets.keySet()) {
+        sequenceNames.put(partitionNum, StringUtils.format("%s_%s", ioConfig.getBaseSequenceName(), partitionNum));
+      }
+
+      // Set up committer.
+      final Supplier<Committer> committerSupplier = new Supplier<Committer>()
+      {
+        @Override
+        public Committer get()
+        {
+          final Map<Integer, Long> snapshot = ImmutableMap.copyOf(nextOffsets);
+
+          return new Committer()
+          {
+            @Override
+            public Object getMetadata()
+            {
+              return ImmutableMap.of(
+                  METADATA_NEXT_PARTITIONS, new KafkaPartitions(
+                      ioConfig.getStartPartitions().getTopic(),
+                      snapshot
+                  )
+              );
+            }
+
+            @Override
+            public void run()
+            {
+              // Do nothing.
+            }
+          };
+        }
+      };
+
+      Set<Integer> assignment = assignPartitionsAndSeekToNext(consumer, topic);
+
+      // Main loop.
+      // Could eventually support leader/follower mode (for keeping replicas more in sync)
+      boolean stillReading = !assignment.isEmpty();
+      status = Status.READING;
+      try {
+        while (stillReading) {
+          if (possiblyPause(assignment)) {
+            // The partition assignments may have changed while paused by a call to setEndOffsets() so reassign
+            // partitions upon resuming. This is safe even if the end offsets have not been modified.
+            assignment = assignPartitionsAndSeekToNext(consumer, topic);
+
+            if (assignment.isEmpty()) {
+              log.info("All partitions have been fully read");
+              publishOnStop.set(true);
+              stopRequested.set(true);
+            }
+          }
+
+          if (stopRequested.get()) {
+            break;
+          }
+
+          // The retrying business is because the KafkaConsumer throws OffsetOutOfRangeException if the seeked-to
+          // offset is not present in the topic-partition. This can happen if we're asking a task to read from data
+          // that has not been written yet (which is totally legitimate). So let's wait for it to show up.
+          ConsumerRecords<byte[], byte[]> records = ConsumerRecords.empty();
+          try {
+            records = consumer.poll(KafkaIndexTask.POLL_TIMEOUT);
+          }
+          catch (OffsetOutOfRangeException e) {
+            log.warn("OffsetOutOfRangeException with message [%s]", e.getMessage());
+            possiblyResetOffsetsOrWait(e.offsetOutOfRangePartitions(), consumer, toolbox);
+            stillReading = ioConfig.isPauseAfterRead() || !assignment.isEmpty();
+          }
+
+          for (ConsumerRecord<byte[], byte[]> record : records) {
+            if (log.isTraceEnabled()) {
+              log.trace(
+                  "Got topic[%s] partition[%d] offset[%,d].",
+                  record.topic(),
+                  record.partition(),
+                  record.offset()
+              );
+            }
+
+            if (record.offset() < endOffsets.get(record.partition())) {
+              if (record.offset() != nextOffsets.get(record.partition())) {
+                if (ioConfig.isSkipOffsetGaps()) {
+                  log.warn(
+                      "Skipped to offset[%,d] after offset[%,d] in partition[%d].",
+                      record.offset(),
+                      nextOffsets.get(record.partition()),
+                      record.partition()
+                  );
+                } else {
+                  throw new ISE(
+                      "WTF?! Got offset[%,d] after offset[%,d] in partition[%d].",
+                      record.offset(),
+                      nextOffsets.get(record.partition()),
+                      record.partition()
+                  );
+                }
+              }
+
+              try {
+                final byte[] valueBytes = record.value();
+                final List<InputRow> rows = valueBytes == null
+                                            ? Utils.nullableListOf((InputRow) null)
+                                            : parser.parseBatch(ByteBuffer.wrap(valueBytes));
+                boolean isPersistRequired = false;
+                final Map<String, Set<SegmentIdentifier>> segmentsToMoveOut = new HashMap<>();
+
+                for (InputRow row : rows) {
+                  if (row != null && task.withinMinMaxRecordTime(row)) {
+                    final String sequenceName = sequenceNames.get(record.partition());
+                    final AppenderatorDriverAddResult addResult = driver.add(
+                        row,
+                        sequenceName,
+                        committerSupplier,
+                        false,
+                        false
+                    );
+
+                    if (addResult.isOk()) {
+                      // If the number of rows in the segment exceeds the threshold after adding a row,
+                      // move the segment out from the active segments of BaseAppenderatorDriver to make a new segment.
+                      if (addResult.getNumRowsInSegment() > tuningConfig.getMaxRowsPerSegment()) {
+                        segmentsToMoveOut.computeIfAbsent(sequenceName, k -> new HashSet<>())
+                                         .add(addResult.getSegmentIdentifier());
+                      }
+                      isPersistRequired |= addResult.isPersistRequired();
+                    } else {
+                      // Failure to allocate segment puts determinism at risk, bail out to be safe.
+                      // May want configurable behavior here at some point.
+                      // If we allow continuing, then consider blacklisting the interval for a while to avoid constant checks.
+                      throw new ISE("Could not allocate segment for row with timestamp[%s]", row.getTimestamp());
+                    }
+
+                    if (addResult.getParseException() != null) {
+                      handleParseException(addResult.getParseException(), record);
+                    } else {
+                      fireDepartmentMetrics.incrementProcessed();
+                    }
+                  } else {
+                    fireDepartmentMetrics.incrementThrownAway();
+                  }
+                }
+
+                if (isPersistRequired) {
+                  driver.persist(committerSupplier.get());
+                }
+                segmentsToMoveOut.entrySet().forEach(sequenceSegments -> driver.moveSegmentOut(
+                    sequenceSegments.getKey(),
+                    sequenceSegments.getValue().stream().collect(Collectors.toList())
+                ));
+              }
+              catch (ParseException e) {
+                handleParseException(e, record);
+              }
+
+              nextOffsets.put(record.partition(), record.offset() + 1);
+            }
+
+            if (nextOffsets.get(record.partition()).equals(endOffsets.get(record.partition()))
+                && assignment.remove(record.partition())) {
+              log.info("Finished reading topic[%s], partition[%,d].", record.topic(), record.partition());
+              KafkaIndexTask.assignPartitions(consumer, topic, assignment);
+              stillReading = ioConfig.isPauseAfterRead() || !assignment.isEmpty();
+            }
+          }
+        }
+        ingestionState = IngestionState.COMPLETED;
+      }
+      catch (Exception e) {
+        log.error(e, "Encountered exception in runLegacy() before persisting.");
+        throw e;
+      }
+      finally {
+        driver.persist(committerSupplier.get()); // persist pending data
+      }
+
+      synchronized (statusLock) {
+        if (stopRequested.get() && !publishOnStop.get()) {
+          throw new InterruptedException("Stopping without publishing");
+        }
+
+        status = Status.PUBLISHING;
+      }
+
+      final TransactionalSegmentPublisher publisher = (segments, commitMetadata) -> {
+        final KafkaPartitions finalPartitions = toolbox.getObjectMapper().convertValue(
+            ((Map) Preconditions.checkNotNull(commitMetadata, "commitMetadata")).get(METADATA_NEXT_PARTITIONS),
+            KafkaPartitions.class
+        );
+
+        // Sanity check, we should only be publishing things that match our desired end state.
+        if (!endOffsets.equals(finalPartitions.getPartitionOffsetMap())) {
+          throw new ISE("WTF?! Driver attempted to publish invalid metadata[%s].", commitMetadata);
+        }
+
+        final SegmentTransactionalInsertAction action;
+
+        if (ioConfig.isUseTransaction()) {
+          action = new SegmentTransactionalInsertAction(
+              segments,
+              new KafkaDataSourceMetadata(ioConfig.getStartPartitions()),
+              new KafkaDataSourceMetadata(finalPartitions)
+          );
+        } else {
+          action = new SegmentTransactionalInsertAction(segments, null, null);
+        }
+
+        log.info("Publishing with isTransaction[%s].", ioConfig.isUseTransaction());
+
+        return toolbox.getTaskActionClient().submit(action).isSuccess();
+      };
+
+      // Supervised kafka tasks are killed by KafkaSupervisor if they are stuck during publishing segments or waiting
+      // for hand off. See KafkaSupervisorIOConfig.completionTimeout.
+      final SegmentsAndMetadata published = driver.publish(
+          publisher,
+          committerSupplier.get(),
+          sequenceNames.values()
+      ).get();
+
+      final Future<SegmentsAndMetadata> handoffFuture = driver.registerHandoff(published);
+      final SegmentsAndMetadata handedOff;
+      if (tuningConfig.getHandoffConditionTimeout() == 0) {
+        handedOff = handoffFuture.get();
+      } else {
+        handedOff = handoffFuture.get(tuningConfig.getHandoffConditionTimeout(), TimeUnit.MILLISECONDS);
+      }
+
+      if (handedOff == null) {
+        throw new ISE("Transaction failure publishing segments, aborting");
+      } else {
+        log.info(
+            "Published segments[%s] with metadata[%s].",
+            Joiner.on(", ").join(
+                Iterables.transform(
+                    handedOff.getSegments(),
+                    new Function<DataSegment, String>()
+                    {
+                      @Override
+                      public String apply(DataSegment input)
+                      {
+                        return input.getIdentifier();
+                      }
+                    }
+                )
+            ),
+            Preconditions.checkNotNull(handedOff.getCommitMetadata(), "commitMetadata")
+        );
+      }
+    }
+    catch (InterruptedException | RejectedExecutionException e) {
+      // handle the InterruptedException that gets wrapped in a RejectedExecutionException
+      if (e instanceof RejectedExecutionException
+          && (e.getCause() == null || !(e.getCause() instanceof InterruptedException))) {
+        throw e;
+      }
+
+      // if we were interrupted because we were asked to stop, handle the exception and return success, else rethrow
+      if (!stopRequested.get()) {
+        Thread.currentThread().interrupt();
+        throw e;
+      }
+
+      log.info("The task was asked to stop before completing");
+    }
+    finally {
+      if (chatHandlerProvider.isPresent()) {
+        chatHandlerProvider.get().unregister(task.getId());
+      }
+
+      toolbox.getDruidNodeAnnouncer().unannounce(discoveryDruidNode);
+      toolbox.getDataSegmentServerAnnouncer().unannounce();
+    }
+
+    toolbox.getTaskReportFileWriter().write(getTaskCompletionReports(null));
+    return TaskStatus.success(
+        task.getId(),
+        null
+    );
+  }
+
+  private Set<Integer> assignPartitionsAndSeekToNext(KafkaConsumer consumer, String topic)
+  {
+    // Initialize consumer assignment.
+    final Set<Integer> assignment = Sets.newHashSet();
+    for (Map.Entry<Integer, Long> entry : nextOffsets.entrySet()) {
+      final long endOffset = endOffsets.get(entry.getKey());
+      if (entry.getValue() < endOffset) {
+        assignment.add(entry.getKey());
+      } else if (entry.getValue() == endOffset) {
+        log.info("Finished reading partition[%d].", entry.getKey());
+      } else {
+        throw new ISE(
+            "WTF?! Cannot start from offset[%,d] > endOffset[%,d]",
+            entry.getValue(),
+            endOffset
+        );
+      }
+    }
+
+    task.assignPartitions(consumer, topic, assignment);
+
+    // Seek to starting offsets.
+    for (final int partition : assignment) {
+      final long offset = nextOffsets.get(partition);
+      log.info("Seeking partition[%d] to offset[%,d].", partition, offset);
+      consumer.seek(new TopicPartition(topic, partition), offset);
+    }
+
+    return assignment;
+  }
+
+  /**
+   * Checks if the pauseRequested flag was set and if so blocks:
+   * a) if pauseMillis == PAUSE_FOREVER, until pauseRequested is cleared
+   * b) if pauseMillis != PAUSE_FOREVER, until pauseMillis elapses -or- pauseRequested is cleared
+   * <p/>
+   * If pauseMillis is changed while paused, the new pause timeout will be applied. This allows adjustment of the
+   * pause timeout (making a timed pause into an indefinite pause and vice versa is valid) without having to resume
+   * and ensures that the loop continues to stay paused without ingesting any new events. You will need to signal
+   * shouldResume after adjusting pauseMillis for the new value to take effect.
+   * <p/>
+   * Sets paused = true and signals paused so callers can be notified when the pause command has been accepted.
+   * <p/>
+   * Additionally, pauses if all partitions assignments have been read and pauseAfterRead flag is set.
+   *
+   * @return true if a pause request was handled, false otherwise
+   */
+  private boolean possiblyPause(Set<Integer> assignment) throws InterruptedException
+  {
+    pauseLock.lockInterruptibly();
+    try {
+      if (ioConfig.isPauseAfterRead() && assignment.isEmpty()) {
+        pauseMillis = KafkaIndexTask.PAUSE_FOREVER;
+        pauseRequested = true;
+      }
+
+      if (pauseRequested) {
+        status = Status.PAUSED;
+        long nanos = 0;
+        hasPaused.signalAll();
+
+        while (pauseRequested) {
+          if (pauseMillis == KafkaIndexTask.PAUSE_FOREVER) {
+            log.info("Pausing ingestion until resumed");
+            shouldResume.await();
+          } else {
+            if (pauseMillis > 0) {
+              log.info("Pausing ingestion for [%,d] ms", pauseMillis);
+              nanos = TimeUnit.MILLISECONDS.toNanos(pauseMillis);
+              pauseMillis = 0;
+            }
+            if (nanos <= 0L) {
+              pauseRequested = false; // timeout elapsed
+            }
+            nanos = shouldResume.awaitNanos(nanos);
+          }
+        }
+
+        status = Status.READING;
+        shouldResume.signalAll();
+        log.info("Ingestion loop resumed");
+        return true;
+      }
+    }
+    finally {
+      pauseLock.unlock();
+    }
+
+    return false;
+  }
+
+  private void possiblyResetOffsetsOrWait(
+      Map<TopicPartition, Long> outOfRangePartitions,
+      KafkaConsumer<byte[], byte[]> consumer,
+      TaskToolbox taskToolbox
+  ) throws InterruptedException, IOException
+  {
+    final Map<TopicPartition, Long> resetPartitions = Maps.newHashMap();
+    boolean doReset = false;
+    if (tuningConfig.isResetOffsetAutomatically()) {
+      for (Map.Entry<TopicPartition, Long> outOfRangePartition : outOfRangePartitions.entrySet()) {
+        final TopicPartition topicPartition = outOfRangePartition.getKey();
+        final long nextOffset = outOfRangePartition.getValue();
+        // seek to the beginning to get the least available offset
+        consumer.seekToBeginning(Collections.singletonList(topicPartition));
+        final long leastAvailableOffset = consumer.position(topicPartition);
+        // reset the seek
+        consumer.seek(topicPartition, nextOffset);
+        // Reset consumer offset if resetOffsetAutomatically is set to true
+        // and the current message offset in the kafka partition is more than the
+        // next message offset that we are trying to fetch
+        if (leastAvailableOffset > nextOffset) {
+          doReset = true;
+          resetPartitions.put(topicPartition, nextOffset);
+        }
+      }
+    }
+
+    if (doReset) {
+      sendResetRequestAndWait(resetPartitions, taskToolbox);
+    } else {
+      log.warn("Retrying in %dms", task.getPollRetryMs());
+      pollRetryLock.lockInterruptibly();
+      try {
+        long nanos = TimeUnit.MILLISECONDS.toNanos(task.getPollRetryMs());
+        while (nanos > 0L && !pauseRequested && !stopRequested.get()) {
+          nanos = isAwaitingRetry.awaitNanos(nanos);
+        }
+      }
+      finally {
+        pollRetryLock.unlock();
+      }
+    }
+  }
+
+  private void sendResetRequestAndWait(Map<TopicPartition, Long> outOfRangePartitions, TaskToolbox taskToolbox)
+      throws IOException
+  {
+    Map<Integer, Long> partitionOffsetMap = Maps.newHashMap();
+    for (Map.Entry<TopicPartition, Long> outOfRangePartition : outOfRangePartitions.entrySet()) {
+      partitionOffsetMap.put(outOfRangePartition.getKey().partition(), outOfRangePartition.getValue());
+    }
+    boolean result = taskToolbox.getTaskActionClient()
+                                .submit(new ResetDataSourceMetadataAction(
+                                    task.getDataSource(),
+                                    new KafkaDataSourceMetadata(new KafkaPartitions(
+                                        ioConfig.getStartPartitions()
+                                                .getTopic(),
+                                        partitionOffsetMap
+                                    ))
+                                ));
+
+    if (result) {
+      log.makeAlert("Resetting Kafka offsets for datasource [%s]", task.getDataSource())
+         .addData("partitions", partitionOffsetMap.keySet())
+         .emit();
+      // wait for being killed by supervisor
+      requestPause(KafkaIndexTask.PAUSE_FOREVER);
+    } else {
+      log.makeAlert("Failed to send reset request for partitions [%s]", partitionOffsetMap.keySet()).emit();
+    }
+  }
+
+  private void requestPause(long pauseMillis)
+  {
+    this.pauseMillis = pauseMillis;
+    pauseRequested = true;
+  }
+
+  private void handleParseException(ParseException pe, ConsumerRecord<byte[], byte[]> record)
+  {
+    if (pe.isFromPartiallyValidRow()) {
+      fireDepartmentMetrics.incrementProcessedWithErrors();
+    } else {
+      fireDepartmentMetrics.incrementUnparseable();
+    }
+
+    if (tuningConfig.isLogParseExceptions()) {
+      log.error(
+          pe,
+          "Encountered parse exception on row from partition[%d] offset[%d]",
+          record.partition(),
+          record.offset()
+      );
+    }
+
+    if (savedParseExceptions != null) {
+      savedParseExceptions.add(pe);
+    }
+
+    if (fireDepartmentMetrics.unparseable() + fireDepartmentMetrics.processedWithErrors()
+        > tuningConfig.getMaxParseExceptions()) {
+      log.error("Max parse exceptions exceeded, terminating task...");
+      throw new RuntimeException("Max parse exceptions exceeded, terminating task...");
+    }
+  }
+
+  private Map<String, TaskReport> getTaskCompletionReports(@Nullable String errorMsg)
+  {
+    return TaskReport.buildTaskReports(
+        new IngestionStatsAndErrorsTaskReport(
+            task.getId(),
+            new IngestionStatsAndErrorsTaskReportData(
+                ingestionState,
+                getTaskCompletionUnparseableEvents(),
+                getTaskCompletionRowStats(),
+                errorMsg
+            )
+        )
+    );
+  }
+
+  private Map<String, Object> getTaskCompletionUnparseableEvents()
+  {
+    Map<String, Object> unparseableEventsMap = Maps.newHashMap();
+    List<String> buildSegmentsParseExceptionMessages = IndexTaskUtils.getMessagesFromSavedParseExceptions(savedParseExceptions);
+    if (buildSegmentsParseExceptionMessages != null) {
+      unparseableEventsMap.put("buildSegments", buildSegmentsParseExceptionMessages);
+    }
+    return unparseableEventsMap;
+  }
+
+  private Map<String, Object> getTaskCompletionRowStats()
+  {
+    Map<String, Object> metrics = Maps.newHashMap();
+    if (metricsGetter != null) {
+      metrics.put(
+          "buildSegments",
+          metricsGetter.getTotalMetrics()
+      );
+    }
+    return metrics;
+  }
+
+  @Override
+  public void stopGracefully()
+  {
+    log.info("Stopping gracefully (status: [%s])", status);
+    stopRequested.set(true);
+
+    synchronized (statusLock) {
+      if (status == Status.PUBLISHING) {
+        runThread.interrupt();
+        return;
+      }
+    }
+
+    try {
+      if (pauseLock.tryLock(KafkaIndexTask.LOCK_ACQUIRE_TIMEOUT_SECONDS, TimeUnit.SECONDS)) {
+        try {
+          if (pauseRequested) {
+            pauseRequested = false;
+            shouldResume.signalAll();
+          }
+        }
+        finally {
+          pauseLock.unlock();
+        }
+      } else {
+        log.warn("While stopping: failed to acquire pauseLock before timeout, interrupting run thread");
+        runThread.interrupt();
+        return;
+      }
+
+      if (pollRetryLock.tryLock(KafkaIndexTask.LOCK_ACQUIRE_TIMEOUT_SECONDS, TimeUnit.SECONDS)) {
+        try {
+          isAwaitingRetry.signalAll();
+        }
+        finally {
+          pollRetryLock.unlock();
+        }
+      } else {
+        log.warn("While stopping: failed to acquire pollRetryLock before timeout, interrupting run thread");
+        runThread.interrupt();
+      }
+    }
+    catch (Exception e) {
+      Throwables.propagate(e);
+    }
+  }
+
+  /**
+   * Authorizes action to be performed on this task's datasource
+   *
+   * @return authorization result
+   */
+  private Access authorizationCheck(final HttpServletRequest req, Action action)
+  {
+    return IndexTaskUtils.datasourceAuthorizationCheck(req, action, task.getDataSource(), authorizerMapper);
+  }
+
+  @POST
+  @Path("/stop")
+  public Response stop(@Context final HttpServletRequest req)
+  {
+    authorizationCheck(req, Action.WRITE);
+    stopGracefully();
+    return Response.status(Response.Status.OK).build();
+  }
+
+  @GET
+  @Path("/status")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Status getStatusHTTP(@Context final HttpServletRequest req)
+  {
+    authorizationCheck(req, Action.READ);
+    return status;
+  }
+
+  @Override
+  public Status getStatus()
+  {
+    return status;
+  }
+
+  @GET
+  @Path("/offsets/current")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Map<Integer, Long> getCurrentOffsets(@Context final HttpServletRequest req)
+  {
+    authorizationCheck(req, Action.READ);
+    return getCurrentOffsets();
+  }
+
+  @Override
+  public Map<Integer, Long> getCurrentOffsets()
+  {
+    return nextOffsets;
+  }
+
+  @GET
+  @Path("/offsets/end")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Map<Integer, Long> getEndOffsetsHTTP(@Context final HttpServletRequest req)
+  {
+    authorizationCheck(req, Action.READ);
+    return getEndOffsets();
+  }
+
+  @Override
+  public Map<Integer, Long> getEndOffsets()
+  {
+    return endOffsets;
+  }
+
+  @Override
+  public Response setEndOffsets(Map<Integer, Long> offsets, boolean resume, boolean finish) throws InterruptedException
+  {
+    // finish is not used in this mode
+    return setEndOffsets(offsets, resume);
+  }
+
+  @POST
+  @Path("/offsets/end")
+  @Consumes(MediaType.APPLICATION_JSON)
+  @Produces(MediaType.APPLICATION_JSON)
+  public Response setEndOffsetsHTTP(
+      Map<Integer, Long> offsets,
+      @QueryParam("resume") @DefaultValue("false") final boolean resume,
+      @Context final HttpServletRequest req
+  ) throws InterruptedException
+  {
+    authorizationCheck(req, Action.WRITE);
+    return setEndOffsets(offsets, resume);
+  }
+
+  @GET
+  @Path("/rowStats")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Response getRowStats(
+      @Context final HttpServletRequest req
+  )
+  {
+    authorizationCheck(req, Action.READ);
+    Map<String, Object> returnMap = Maps.newHashMap();
+    Map<String, Object> totalsMap = Maps.newHashMap();
+
+    if (metricsGetter != null) {
+      totalsMap.put(
+          "buildSegments",
+          metricsGetter.getTotalMetrics()
+      );
+    }
+
+    returnMap.put("totals", totalsMap);
+    return Response.ok(returnMap).build();
+  }
+
+  @GET
+  @Path("/unparseableEvents")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Response getUnparseableEvents(
+      @Context final HttpServletRequest req
+  )
+  {
+    authorizationCheck(req, Action.READ);
+    List<String> events = IndexTaskUtils.getMessagesFromSavedParseExceptions(savedParseExceptions);
+    return Response.ok(events).build();
+  }
+
+  public Response setEndOffsets(
+      Map<Integer, Long> offsets,
+      final boolean resume
+  ) throws InterruptedException
+  {
+    if (offsets == null) {
+      return Response.status(Response.Status.BAD_REQUEST)
+                     .entity("Request body must contain a map of { partition:endOffset }")
+                     .build();
+    } else if (!endOffsets.keySet().containsAll(offsets.keySet())) {
+      return Response.status(Response.Status.BAD_REQUEST)
+                     .entity(
+                         StringUtils.format(
+                             "Request contains partitions not being handled by this task, my partitions: %s",
+                             endOffsets.keySet()
+                         )
+                     )
+                     .build();
+    }
+
+    pauseLock.lockInterruptibly();
+    try {
+      if (!isPaused()) {
+        return Response.status(Response.Status.BAD_REQUEST)
+                       .entity("Task must be paused before changing the end offsets")
+                       .build();
+      }
+
+      for (Map.Entry<Integer, Long> entry : offsets.entrySet()) {
+        if (entry.getValue().compareTo(nextOffsets.get(entry.getKey())) < 0) {
+          return Response.status(Response.Status.BAD_REQUEST)
+                         .entity(
+                             StringUtils.format(
+                                 "End offset must be >= current offset for partition [%s] (current: %s)",
+                                 entry.getKey(),
+                                 nextOffsets.get(entry.getKey())
+                             )
+                         )
+                         .build();
+        }
+      }
+
+      endOffsets.putAll(offsets);
+      log.info("endOffsets changed to %s", endOffsets);
+    }
+    finally {
+      pauseLock.unlock();
+    }
+
+    if (resume) {
+      resume();
+    }
+
+    return Response.ok(endOffsets).build();
+  }
+
+  private boolean isPaused()
+  {
+    return status == Status.PAUSED;
+  }
+
+  /**
+   * Signals the ingestion loop to pause.
+   *
+   * @param timeout how long to pause for before resuming in milliseconds, <= 0 means indefinitely
+   *
+   * @return one of the following Responses: 400 Bad Request if the task has started publishing; 202 Accepted if the
+   * method has timed out and returned before the task has paused; 200 OK with a map of the current partition offsets
+   * in the response body if the task successfully paused
+   */
+  @POST
+  @Path("/pause")
+  @Produces(MediaType.APPLICATION_JSON)
+  public Response pauseHTTP(
+      @QueryParam("timeout") @DefaultValue("0") final long timeout,
+      @Context final HttpServletRequest req
+  ) throws InterruptedException
+  {
+    authorizationCheck(req, Action.WRITE);
+    return pause(timeout);
+  }
+
+  @Override
+  public Response pause(final long timeout) throws InterruptedException
+  {
+    if (!(status == Status.PAUSED || status == Status.READING)) {
+      return Response.status(Response.Status.BAD_REQUEST)
+                     .entity(StringUtils.format("Can't pause, task is not in a pausable state (state: [%s])", status))
+                     .build();
+    }
+
+    pauseLock.lockInterruptibly();
+    try {
+      pauseMillis = timeout <= 0 ? KafkaIndexTask.PAUSE_FOREVER : timeout;
+      pauseRequested = true;
+
+      pollRetryLock.lockInterruptibly();
+      try {
+        isAwaitingRetry.signalAll();
+      }
+      finally {
+        pollRetryLock.unlock();
+      }
+
+      if (isPaused()) {
+        shouldResume.signalAll(); // kick the monitor so it re-awaits with the new pauseMillis
+      }
+
+      long nanos = TimeUnit.SECONDS.toNanos(2);
+      while (!isPaused()) {
+        if (nanos <= 0L) {
+          return Response.status(Response.Status.ACCEPTED)
+                         .entity("Request accepted but task has not yet paused")
+                         .build();
+        }
+        nanos = hasPaused.awaitNanos(nanos);
+      }
+    }
+    finally {
+      pauseLock.unlock();
+    }
+
+    try {
+      return Response.ok().entity(objectMapper.writeValueAsString(getCurrentOffsets())).build();
+    }
+    catch (JsonProcessingException e) {
+      throw Throwables.propagate(e);
+    }
+  }
+
+  @POST
+  @Path("/resume")
+  public Response resumeHTTP(@Context final HttpServletRequest req) throws InterruptedException
+  {
+    authorizationCheck(req, Action.WRITE);
+    resume();
+    return Response.status(Response.Status.OK).build();
+  }
+
+  @Override
+  public void resume() throws InterruptedException
+  {
+    pauseLock.lockInterruptibly();
+    try {
+      pauseRequested = false;
+      shouldResume.signalAll();
+
+      long nanos = TimeUnit.SECONDS.toNanos(5);
+      while (isPaused()) {
+        if (nanos <= 0L) {
+          throw new RuntimeException("Resume command was not accepted within 5 seconds");
+        }
+        nanos = shouldResume.awaitNanos(nanos);
+      }
+    }
+    finally {
+      pauseLock.unlock();
+    }
+  }
+
+  @GET
+  @Path("/time/start")
+  @Produces(MediaType.APPLICATION_JSON)
+  public DateTime getStartTime(@Context final HttpServletRequest req)
+  {
+    authorizationCheck(req, Action.WRITE);
+    return startTime;
+  }
+}
